fix(services): disallow creation of certain types
The ServiceFactory.create method has had an additional check added to it
in an effort to ensure that only custom class types are passed to it. If
a type is passed to it which is not a custom class type, this is most
likely an error, so we don't want to silently allow it.